Address reviewer comment
